fix(local-resolver): rename file to match class name
The LocalResolver's filename wasn't match the class name.
